refatorando o codigo
