+++ src/data/FeedDatabase.ts
@@ -1,7 +1,8 @@
 import { BaseDatabase } from "./BaseDatabase";
+import { Post } from "../model/Post";
 
 export class FeedDatabase extends BaseDatabase {
-    public async createFeed(userId: string): Promise<any> {
+    public async createFeed(userId: string): Promise<Post []> {
         const result = await this.getConnection().raw(`
             SELECT Posts.posts_id, photo, description, creation_date, type, User.id, User.name
             FROM Posts
@@ -14,5 +15,4 @@ export class FeedDatabase extends BaseDatabase {
         return result[0]
     }
 
-    // public async getAllPosts(): Promise<any> {}
 };
