+++ src/endpoints/login.ts
@@ -1,50 +1,51 @@
-import { Request, Response } from "express";
-import HashManager from "../services/HashManager";
-import { Authenticator } from "../services/Authenticator";
-import { BaseDatabase } from "../data/BaseDatabase";
-import { UserDatabase } from "../data/UserDatabase";
+// import { Request, Response } from "express";
+// import HashManager from "../services/HashManager";
+// import { Authenticator } from "../services/Authenticator";
+// import { BaseDatabase } from "../data/BaseDatabase";
+// import { UserDatabase } from "../data/UserDataBase";
 
-/* Login de usuário criado - POST => /login */
-export const loginEndpoint = async (req: Request, res: Response) => {
-    try {
-        const email = req.body.email;
-        const password = req.body.password;
-        const role = req.body.role;
 
-        const userDataBase = new UserDatabase();
-        const user = await userDataBase.getUserByEmail(email);
+// /* Login de usuário criado - POST => /login */
+// export const loginEndpoint = async (req: Request, res: Response) => {
+//     try {
+//         const email = req.body.email;
+//         const password = req.body.password;
+//         const role = req.body.role;
 
-        const hashManager = new HashManager();
+//         const userDataBase = new UserDatabase();
+//         const user = await userDataBase.getUserByEmail(email);
 
-        /* Validação senha */
-        const isPasswordCorrect = await hashManager.compare(password, user.password);
+//         const hashManager = new HashManager();
 
-        if (!isPasswordCorrect) {
-            throw new Error("Usuário ou senha inválidos")
-        };
+//         /* Validação senha */
+//         const isPasswordCorrect = await hashManager.compare(password, user.password);
 
-        /* Validação email */
-        if (!email || email.indexOf("@") === -1) {
-            throw new Error("Email inválido")
-        };
+//         if (!isPasswordCorrect) {
+//             throw new Error("Usuário ou senha inválidos")
+//         };
 
-        /* Validação campos */
-        if (!email || !password) {
-            throw new Error("Parâmetros inválidos")
-        };
+//         /* Validação email */
+//         if (!email || email.indexOf("@") === -1) {
+//             throw new Error("Email inválido")
+//         };
 
-        const authenticator = new Authenticator();
+//         /* Validação campos */
+//         if (!email || !password) {
+//             throw new Error("Parâmetros inválidos")
+//         };
 
-        const token = authenticator.generateToken({
-            id: user.id,
-            role: role
-        });
+//         const authenticator = new Authenticator();
 
-        res.status(200).send({ access_token: token });
+//         const token = authenticator.generateToken({
+//             id: user.id,
+//             role: role
+//         });
 
-    } catch (err) {
-        res.status(400).send({ message: err.message });
-    };
+//         res.status(200).send({ access_token: token });
 
-    await BaseDatabase.destroyConnection();
-};
+//     } catch (err) {
+//         res.status(400).send({ message: err.message });
+//     };
+
+//     await BaseDatabase.destroyConnection();
+// };
