+++ src/index.ts
@@ -1,7 +1,9 @@
 import express from "express";
 import dotenv from "dotenv";
 import { AddressInfo } from "net";
-import { routes } from "./routes";
+import { userRouter } from "./routes/userRoutes";
+import { feedRouter } from "./routes/feedRoutes";
+import { friendRouter } from "./routes/friendRoutes";
 
 dotenv.config();
 
@@ -9,7 +11,9 @@ const app = express();
 
 app.use(express.json());
 
-app.use(routes);
+app.use("/user", userRouter);
+app.use("/feed", feedRouter);
+app.use("/friend", friendRouter);
 
 /* Configurações do express para iniciar o servidor */
 const server = app.listen(process.env.PORT || 3003, () => {
