+++ src/routes/feedRoutes.ts
@@ -1,0 +1,9 @@
+import express from "express";
+import { FeedController } from "../controller/FeedController";
+
+export const feedRouter = express.Router();
+
+const feedController = new FeedController();
+
+feedRouter.post("/post", feedController.feedEndpoint);
+feedRouter.post("/create", feedController.createPostEndpoint);
