+++ src/routes/userRoutes.ts
@@ -1,3 +1,9 @@
-export class userRoutes {
-    
-};
+import express from "express";
+import { UserController } from "../controller/UserController";
+
+export const userRouter = express.Router();
+
+const userController = new UserController();
+
+userRouter.post("/signup", userController.signUpEndpoint);
+userRouter.post("/login", userController.loginEndpoint);
